chore: status bar wip
