+++ components/status-bar/status-bar-controller.ts
