+++ components/status-bar/status-bar-view.ts
